+++ node/directives/validateStoreUserAccess.ts
@@ -15,6 +15,7 @@ import {
 export class ValidateStoreUserAccess extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
     const { resolve = defaultFieldResolver } = field
+    const { orgPermission } = this.args
 
     field.resolve = async (
       root: any,
@@ -45,7 +46,8 @@ export class ValidateStoreUserAccess extends SchemaDirectiveVisitor {
       const { hasAdminToken, hasValidAdminToken } = await validateAdminToken(
         context,
         adminUserAuthToken as string,
-        metricFields
+        metricFields,
+        orgPermission
       )
 
       // add admin token metrics