+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added optional argument to validateAdminUserAccess and validateStoreUserAccess directives to check specific user permissions via LM API.
+
 ## [1.44.6] - 2024-09-05
 
 ### Fixed